app/requirements.txt
django==4.0.4
uwsgi==2.0.20
psycopg2-binary==2.9
faker==19.6.0
flake8==4.0
python-dotenv==1.0.0
django-split-settings==1.2.0
django-debug-toolbar==4.2.0
pytest==7.4.2
psycopg2-binary==2.9
faker==19.6.0
flake8==4.0
python-dotenv==1.0.0
django-split-settings==1.2.0
django-debug-toolbar==4.2.0
pytest==7.4.2
django-cors-headers==4.3.1
requests==2.31.0
yarl==1.9.4
